grand  fix

+++ src/exe/client_server_template.hpp
@@ -26,7 +26,7 @@ void Test() {
     {
         std::random_device rd;
         std::mt19937 gen(rd());
-        std::uniform_int_distribution<uint64_t> idDistrib(1);
+        std::uniform_int_distribution<uint32_t> idDistrib(1);
         std::uniform_int_distribution<uint32_t> valueDistrib(0);
         initCells.reserve(cellCount);
         for (uint64_t idx = 0; idx < cellCount; ++idx) {
